requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.40.0
tokenizers>=0.15.2
pennylane>=0.35.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
tqdm>=4.66
Pillow>=10.0
